remove argument handling class from common module
RawDescriptionDefaultHelpFormatter was copied to the
analyzer and web parts.
Additionally a new arg module was created for the analyzer part to have
argument parsing related classes and function there,
OrderedCheckersAction was moved there too.
