+++ analyzer/codechecker_analyzer/arg.py
@@ -1,0 +1,53 @@
+# -------------------------------------------------------------------------
+#                     The CodeChecker Infrastructure
+#   This file is distributed under the University of Illinois Open Source
+#   License. See LICENSE.TXT for details.
+# -------------------------------------------------------------------------
+"""Argument parsing related helper classes and functions."""
+
+from __future__ import print_function
+from __future__ import division
+from __future__ import absolute_import
+
+import argparse
+
+
+class RawDescriptionDefaultHelpFormatter(
+        argparse.RawDescriptionHelpFormatter,
+        argparse.ArgumentDefaultsHelpFormatter
+):
+    """
+    Adds default values to argument help and retains any formatting in
+    descriptions.
+    """
+    pass
+
+
+class OrderedCheckersAction(argparse.Action):
+    """
+    Action to store enabled and disabled checkers
+    and keep ordering from command line.
+
+    Create separate lists based on the checker names for
+    each analyzer.
+    """
+
+    # Users can supply invocation to 'codechecker-analyze' as follows:
+    # -e core -d core.uninitialized -e core.uninitialized.Assign
+    # We must support having multiple '-e' and '-d' options and the order
+    # specified must be kept when the list of checkers are assembled for Clang.
+
+    def __init__(self, option_strings, dest, nargs=None, **kwargs):
+        if nargs is not None:
+            raise ValueError("nargs not allowed")
+        super(OrderedCheckersAction, self).__init__(option_strings, dest,
+                                                    **kwargs)
+
+    def __call__(self, parser, namespace, value, option_string=None):
+
+        if 'ordered_checkers' not in namespace:
+            namespace.ordered_checkers = []
+        ordered_checkers = namespace.ordered_checkers
+        ordered_checkers.append((value, self.dest == 'enable'))
+
+        namespace.ordered_checkers = ordered_checkers
